Testing Python integration
When you click the button start the widget. You write the command to
create an object (now only addBox and addCyl) with the parameters and
then click in OK.
The widget only close when click in CANCEL.
Important to know: when OK is cliked run the command given but also
clear the block pad.